% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_idx_class}
\alias{add_idx_class}
\title{Add class to index according to interval.}
\usage{
add_idx_class(index, itv)
}
\arguments{
\item{index}{A Date or POSIXct object.}

\item{itv}{A character, like "1M".}
}
\value{
A date related class object.
}
\description{
Add class to index according to interval.
}
\examples{
\dontrun{
add_idx_class(df$date, "1M")
}
}
