% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ts.R
\name{cal_factory_ts}
\alias{cal_factory_ts}
\title{Function factory for calculation utilizing ts class.}
\usage{
cal_factory_ts(fun_num, fun_idx, fun_itv)
}
\arguments{
\item{fun_num}{A function to transform a numeric vector.}

\item{fun_idx}{A function to transform a numeric vector representing time.}

\item{fun_itv}{A function to transform a string representing interval.}

\item{...}{Arguments for fun_num.}
}
\value{
A function to transform a tsibble into a tsibble or a tibble.
}
\description{
Function factory for calculation utilizing ts class.
}
\examples{
\dontrun{
tq_log <- cal_factory(
function(num) log(num),
function(idx) idx,
function(itv) itv
)
}
}
