% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_zoo.R
\name{cal_factory_zoo}
\alias{cal_factory_zoo}
\title{Function factory for calculation zoo class.}
\usage{
cal_factory_zoo(fun_num, fun_idx, fun_itv)
}
\arguments{
\item{fun_num}{A function to transform a numeric vector or matrix of zoo class.}

\item{fun_idx}{A function to transform a numeric vector representing time.}

\item{fun_itv}{A function to transform a string representing interval.}

\item{df_ts}{A tbl_ts object.}

\item{...}{Arguments for fun_num.}
}
\value{
A function to transform a tsibble into a tsibble or a tibble.
}
\description{
Function factory for calculation zoo class.
}
\examples{
\dontrun{
tq_rollmean <- cal_factory_zoo(
function(num) zoo::rollmean(num, ...),
function(idx) idx,
function(itv) itv
)
}
}
