% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_time_wise.R
\name{cal_time_wise}
\alias{cal_time_wise}
\title{Function factory for adapting time-wise vector function to tsibble.}
\usage{
cal_time_wise(fun_num)
}
\arguments{
\item{fun_num}{A function to transform a time-wise numeric vector.}
}
\value{
A function to transform a tsibble.
}
\description{
Function factory for adapting time-wise vector function to tsibble.
}
\examples{
\dontrun{
tq_diff <- cal_time_wise(tsibble::difference)
}
}
