% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{change_idx_class_for_xts}
\alias{change_idx_class_for_xts}
\title{Change index class compatible with xts.}
\usage{
change_idx_class_for_xts(index, itv)
}
\arguments{
\item{index}{A Date or POSIXct object.}

\item{itv}{A character, like "1M".}
}
\value{
A date related class object compatible with xts.
}
\description{
Change index class compatible with xts.
}
\examples{
\dontrun{
change_idx_class_for_xts(df$date, "1M")
}
}
