% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eer.R
\docType{data}
\name{eer}
\alias{eer}
\title{Effective exchange rates.}
\format{
Data frame with columns
\describe{
\item{date}{months}
\item{symbol}{"reer" (real), "neer" (nominal) or "deflator"}
\item{Algeria}{areas}
}
}
\source{
https://www.bis.org/statistics/eer.htm
}
\usage{
eer
}
\description{
Monthly Effective exchange rates for 60 areas.
}
\examples{
  eer
}
\keyword{datasets}
