% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tq_gr.R
\name{growth_rate}
\alias{growth_rate}
\title{Growth rate}
\usage{
growth_rate(x, n = 12, annualize = 1, pct = TRUE, order_by = NULL)
}
\arguments{
\item{x}{A numeric vector.}

\item{n}{A positive integer indicating lag.}

\item{annualize}{A positive integer indicating the number to annualize.}

\item{pct}{A logical. If TRUE (default), returns percent growth rate.}

\item{order_by}{A vector to order by.}
}
\value{
A numeric vector of the same length as `x`.
}
\description{
Growth rate
}
\examples{
\dontrun{
# example from base
growth_rate(1:10, 4)
growth_rate(1:10, 1, annualize = 4)
# Use order_by if data not already ordered (example from dplyr)
library(dplyr, warn.conflicts = FALSE)
tsbl <- tsibble(year = 2000:2005, value = (0:5)^2, index = year)
scrambled <- tsbl \%>\% slice(sample(nrow(tsbl)))

wrong <- mutate(scrambled, gr = growth_rate(value))
arrange(wrong, year)

right <- mutate(scrambled, gr = growth_rate(value, order_by = year))
arrange(right, year)
}
}
