% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{itv2list}
\alias{itv2list}
\title{Convert interval to list.}
\usage{
itv2list(itv)
}
\arguments{
\item{itv}{A character, like "1M".}
}
\value{
A list.
}
\description{
Convert interval to list.
}
\examples{
\dontrun{
itv2list("1M")
}
}
