% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tq_ma.R
\name{moving_average}
\alias{moving_average}
\title{Moving average}
\usage{
moving_average(
  x,
  n = 3,
  na.rm = FALSE,
  .align = "right",
  .step = 1L,
  .complete = TRUE,
  order_by = NULL
)
}
\arguments{
\item{x}{A numeric vector.}

\item{n}{A positive integer indicating window width.}

\item{na.rm}{A logical. TRUE or FALSE (default).}

\item{.align}{A position string:
\itemize{
 \item "right" (default)
 \item "left"
 \item "center", if n is odd
 \item "center-right" or "center-left", if n is even
}}

\item{.step}{A positive integer indicating the number of elements to shift. 1L (default).}

\item{.complete}{A logical. If TRUE (default), partial computations are not allowed.}

\item{order_by}{A vector to order by.}
}
\value{
A numeric vector of the same length as `x`.
}
\description{
Moving average
}
\examples{
\dontrun{
# example from base
moving_average(1:10, 3)
moving_average(c(1:5, NA, 7:10), 3, na.rm = TRUE)
moving_average(1:10, 3, .align = "left")
moving_average(1:10, 3, .step = 2)
moving_average(1:10, 3, .complete = FALSE)
# Use order_by if data not already ordered (example from dplyr)
library(dplyr, warn.conflicts = FALSE)
tsbl <- tsibble(year = 2000:2005, value = (0:5)^2, index = year)
scrambled <- tsbl \%>\% slice(sample(nrow(tsbl)))

wrong <- mutate(scrambled, ma = moving_average(value))
arrange(wrong, year)

right <- mutate(scrambled, ma = moving_average(value, order_by = year))
arrange(right, year)
}
}
\seealso{
[slider::slide_dbl]
}
