% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tq_sa.R
\name{season_adjust}
\alias{season_adjust}
\title{Seasonally adjusted}
\usage{
season_adjust(x, order_by, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{order_by}{A date vector to order by.}

\item{...}{Parameters in [seasonal::seas] function.}
}
\value{
A numeric vector of the same length as `x`.
}
\description{
Seasonally adjusted
}
\examples{
\dontrun{
# example from base
season_adjust((0:107)^2, seq(as.Date("2001-01-01"), as.Date("2009-12-01"), "months"))
# Automatically order even if data not already ordered (example from dplyr)
library(dplyr, warn.conflicts = FALSE)
tsbl <- tsibble(month = seq(as.Date("2001-01-01"), as.Date("2009-12-01"), "months"),
 value = (0:107)^2, index = month)
scrambled <- tsbl \%>\% slice(sample(nrow(tsbl)))

right <- mutate(scrambled, sa = season_adjust(value, month))
arrange(right, month)
}
}
\seealso{
[seasonal::seas]
}
