% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_time_wise.R
\name{tq_diff}
\alias{tq_diff}
\title{Calculate differences.}
\usage{
tq_diff(df_ts, ...)
}
\arguments{
\item{df_ts}{A tbl_ts object.}

\item{...}{Parameters in [tsibble::difference] function.}
}
\value{
A tbl_ts object.
}
\description{
Calculate differences.
}
\examples{
\dontrun{
tq_diff(df_ts) # default lag = 1
tq_diff(df_ts, lag = 12)
}
}
\seealso{
[tsibble::difference]
}
