% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{tq_gr}
\alias{tq_gr}
\title{Calculate growth rates.}
\usage{
tq_gr(df_ts, ...)
}
\arguments{
\item{df}{A tbl_ts object.}

\item{n}{Degree of lag for growth.}

\item{annualize}{A numeric value of annual frequency.}
}
\value{
A tbl_ts object.
}
\description{
Calculate growth rates.
}
\examples{
\dontrun{
tq_gr(df)
tq_gr(df, n = 1, annualize = 4)
}
}
