% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ts.R
\name{tq_sa}
\alias{tq_sa}
\title{Calculate seasonally adjusted values.}
\usage{
tq_sa(df_ts, ...)
}
\arguments{
\item{...}{Parameters of seasonal::seas().}

\item{df}{A tbl_ts object.}
}
\value{
A tbl_ts object.
}
\description{
Calculate seasonally adjusted values.
}
\examples{
tq_sa(df)
}
