% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{tq_diff}
\alias{tq_diff}
\title{Calculate differences.}
\usage{
tq_diff(df_ts, ...)
}
\arguments{
\item{df}{A tbl_ts object.}

\item{n}{Degree of lag for difference.}
}
\value{
A tbl_ts object.
}
\description{
Calculate differences.
}
\examples{
\dontrun{
tq_diff(df)
tq_diff(df, n = 12)
}
}
