% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{tq_diff}
\alias{tq_diff}
\title{Calculate differences.}
\usage{
tq_diff(df_ts, ...)
}
\arguments{
\item{df_ts}{A tbl_ts object.}

\item{n}{An interval of difference.}
}
\value{
A tbl_ts object.
}
\description{
Calculate differences.
}
\examples{
\dontrun{
tq_diff(df_ts) # default n = 1
tq_diff(df_ts, n = 12)
}
}
