% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{tq_gr}
\alias{tq_gr}
\title{Calculate growth rates.}
\usage{
tq_gr(df_ts, ...)
}
\arguments{
\item{df_ts}{A tbl_ts object.}

\item{...}{Parameters: n, an interval of difference. annualize, a parameter of power.}
}
\value{
A tbl_ts object.
}
\description{
Calculate growth rates.
}
\examples{
\dontrun{
tq_gr(df_ts) # default n = 12, annualize = 1
tq_gr(df_ts, n = 1, annualize = 4) # annualize quarterly growth rates
}
}
