% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tq_ma.R
\name{tq_ma}
\alias{tq_ma}
\title{Calculate moving averages.}
\usage{
tq_ma(df_ts, ...)
}
\arguments{
\item{df_ts}{A tbl_ts object.}

\item{...}{Parameters in moving_average function.}
}
\value{
A tbl_ts object.
}
\description{
Calculate moving averages.
}
\examples{
\dontrun{
tq_ma(df_ts) # default n = 3, na.rm = FALSE, .align = "right", .step = 1L, .complete = TRUE
tq_ma(df_ts, n = 6, na.rm = TRUE, .align = "center-left")
}
}
