% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{tq_ma}
\alias{tq_ma}
\title{Calculate moving averages.}
\usage{
tq_ma(df_ts, ...)
}
\arguments{
\item{df_ts}{A tbl_ts object.}

\item{...}{Parameters for fun_num.}
}
\value{
A tbl_ts object.
}
\description{
Calculate moving averages.
}
\examples{
\dontrun{
tq_ma(df)
tq_ma(df, n = 6, na.rm = TRUE, .align = "center-left")
}
}
