% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{tq_ma}
\alias{tq_ma}
\title{Calculate moving averages.}
\usage{
tq_ma(df_ts, ...)
}
\arguments{
\item{...}{Arguments for slide_dbl function.}

\item{df}{A tbl_ts object.}

\item{n}{Span of moving average. .size for slide_dbl function.}

\item{na.rm}{An argument for mean function.}

\item{.align}{An argument for slide_dbl function.}
}
\value{
A tbl_ts object.
}
\description{
Calculate moving averages.
}
\examples{
\dontrun{
tq_ma(df)
tq_ma(df, n = 6, na.rm = TRUE, .align = "center-left")
}
}
